Utils: Fix implicit narrowing conversion

+++ utils.c
@@ -120,7 +120,7 @@ void append_to_str_lower(char *str, size_t str_size, const char *src) {
 		return;
 	}
 
-	int i = dest - str;
+	size_t i = dest - str;
 	for (int j = 0; (i < str_size) && (j < strlen(src)); ++i, ++j) {
 		str[i] = tolower(src[j]);
 	}
